04.11 22 提高ADC精度,增加滑动平均滤波 更改四元数的姿态误差计算法 修正内环比例控制的增益问题

+++ Inc/adc.h
@@ -39,7 +39,7 @@ extern ADC_HandleTypeDef hadc1;
 void MX_ADC1_Init(void);
 
 /* USER CODE BEGIN Prototypes */
-double Get_Battery_Voltage(void);
+u16 Get_Battery_Voltage(void);
 
 /* USER CODE END Prototypes */
